+++ litellm/proxy/client/cli/commands/models.py
