--- tests/litellm/proxy/client/cli/groups/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the LiteLLM Proxy Client CLI command groups."""
